Mostly smaller fixes
Changes driven by dotnet's CodeGenBringUpTests (https://github.com/dotnet/runtime/tree/v5.0.6/src/tests/JIT/CodeGenBringUpTests)

+++ language/src/main/java/com/vztekoverflow/bacil/runtime/types/builtin/BuiltinTypes.java
@@ -43,7 +43,7 @@ public class BuiltinTypes {
         int32Type = corlib.findLocalType("System", "Int32");
         uInt32Type = corlib.findLocalType("System", "UInt32");
         int64Type = corlib.findLocalType("System", "Int64");
-        uInt64Type = corlib.findLocalType("System", "UInt16");
+        uInt64Type = corlib.findLocalType("System", "UInt64");
         intPtrType = corlib.findLocalType("System", "IntPtr");
         uIntPtrType = corlib.findLocalType("System", "UIntPtr");
         typedReferenceType = corlib.findLocalType("System", "TypedReference");
